added RESIZE_TRACKING setting

+++ flask_resize/__init__.py
@@ -552,7 +552,8 @@ def resize(image_url, dimensions, format=None, quality=80, fill=False,
                                          'to be set.')
 
     # check if image has already been rendered and if it has changed since
-    if _image_changed(original_path, full_cache_path):
+    tracking_enabled = current_app.config['RESIZE_TRACKING']
+    if tracking_enabled and _image_changed(original_path, full_cache_path):
         os.remove(full_cache_path)
 
     if not os.path.exists(full_cache_path):
@@ -562,9 +563,10 @@ def resize(image_url, dimensions, format=None, quality=80, fill=False,
                        anchor=anchor, quality=quality, progressive=progressive,
                        placeholder_reason=placeholder_reason)
 
-        _add_image_to_catalog(original_path, full_cache_path)
+        if tracking_enabled:
+            _add_image_to_catalog(original_path, full_cache_path)
 
-    if original_path not in image_catalog:
+    if tracking_enabled and original_path not in image_catalog:
         _add_image_to_catalog(original_path, full_cache_path)
 
     return full_cache_url
@@ -660,6 +662,7 @@ class Resize(object):
         app.config.setdefault('RESIZE_CACHE_DIR', 'cache')
         app.config.setdefault('RESIZE_HASH_FILENAME', True)
         app.config.setdefault('RESIZE_HASH_METHOD', 'md5')
+        app.config.setdefault('RESIZE_TRACKING', True)
 
         if app.config['RESIZE_NOOP']:
             return  # No RESIZE_URL or RESIZE_ROOT need to be specified.
@@ -677,30 +680,31 @@ class Resize(object):
         if not resize_root.endswith('/'):
             app.config['RESIZE_ROOT'] = resize_root + '/'
 
-        # load image catalog if exists
-        global image_catalog
-        global catalog_path
-        catalog_path = os.path.join(
-            resize_root, app.config['RESIZE_CACHE_DIR'], CATALOG_FILE)
-        if os.path.isfile(catalog_path):
-            with open(catalog_path, 'rb') as f:
-                try:
-                    image_catalog = pickle.load(f)
-                except TypeError:
-                    logger.error('Error reading image catalog.')
-                    image_catalog = {}
+        # if tracking is enabled load image catalog if exists
+        if app.config['RESIZE_TRACKING']:
+            global image_catalog
+            global catalog_path
+            catalog_path = os.path.join(
+                resize_root, app.config['RESIZE_CACHE_DIR'], CATALOG_FILE)
+            if os.path.isfile(catalog_path):
+                with open(catalog_path, 'rb') as f:
+                    try:
+                        image_catalog = pickle.load(f)
+                    except TypeError:
+                        logger.error('Error reading image catalog.')
+                        image_catalog = {}
 
-                # check if pickled file is a dict
-                catalog_valid = isinstance(image_catalog, dict)
+                    # check if pickled file is a dict
+                    catalog_valid = isinstance(image_catalog, dict)
 
-                # check each item is an ImageCatalogItem
-                if catalog_valid:
-                    for _, item in image_catalog.items():
-                        if not isinstance(item, ImageCatalogItem):
-                            catalog_valid = False
-                            break
+                    # check each item is an ImageCatalogItem
+                    if catalog_valid:
+                        for _, item in image_catalog.items():
+                            if not isinstance(item, ImageCatalogItem):
+                                catalog_valid = False
+                                break
 
-                # if image catalog from pickled file is not valid
-                # reset the catalog
-                if not catalog_valid:
-                    image_catalog = {}
+                    # if image catalog from pickled file is not valid
+                    # reset the catalog
+                    if not catalog_valid:
+                        image_catalog = {}
